# openapi.yaml
openapi: 3.0.3
info:
  title: Cozy API
  description: |
    REST API for Cozy (Bishkek shoe shop) — covers `/api/v1/*` (public storefront,
    used by the Flutter app and the site's AJAX/HTMX calls) and `/admin/api/*`
    (staff-only admin panel), per §6 of `docs/Cozy — Технический ТЗ для
    разработчика.md`.

    This file documents only endpoints that actually exist in the code today —
    don't let it drift from the real route table (`internal/httpapi`,
    `internal/staff`, `internal/points`, `internal/media`).

    Online card payment (`payment_method: online_card`) goes through
    `internal/payments`. Until Bakai Bank grants API access the server runs
    with `PAYMENTS_PROVIDER=mock` (default outside `APP_ENV=prod`): the
    returned `payment_url` is a local test page that simulates the bank. With
    `PAYMENTS_PROVIDER=bakai` (the prod default) online orders are refused with
    `503 payments_not_configured` until the real integration lands.
  version: "0.6.0"
servers:
  - url: http://localhost:8080
    description: Local dev (see cozy_backend/README.md)
...
security: []
paths:
...
  /api/v1/favorites:
    get:
      tags: [Customer]
      summary: List favorited products
      description: |
        Returns full product objects, newest favorite first — each item has
        exactly the `GET /api/v1/products/{id}` shape (`images`, `variants`
        with per-point `stock`), so the app renders the list and opens a
        product without a request per item. Served with a constant number of
        queries regardless of list length. A product that's since gone
        inactive/deleted is silently skipped rather than failing the request.
      security: [{ customerBearerAuth: [] }]
      responses:
        "200":
          description: Favorited products.
          content:
            application/json:
              schema:
                type: object
                properties: { items: { type: array, items: { $ref: "#/components/schemas/ProductDetailResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/cart:
    get:
      tags: [Customer]
      summary: Get the customer's server-side cart
      description: |
        Persisted server-side (not localStorage) so it's shared between the
        website and the mobile app for the same logged-in customer. Lines of
        deactivated products or with more pairs than any one store holds are
        still returned, with `available: false` — checkout would reject
        them with an error naming the product.
      security: [{ customerBearerAuth: [] }]
      responses:
        "200":
          description: Cart lines, enriched with the variant/product details needed to render a cart screen from this one response.
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/CartLine" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    staffSessionCookie:
      type: apiKey
      in: cookie
      name: staff_session
      description: Server-side staff session, set by `POST /admin/api/login`.
    webhookToken:
      type: apiKey
      in: header
      name: X-Webhook-Token
      description: Shared secret for payment provider callbacks (`BAKAI_WEBHOOK_TOKEN`).
    customerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Customer access token from `POST /api/v1/auth/otp/verify` or
        `/api/v1/auth/refresh`. Required on every `Customer`-tagged endpoint
        (favorites, addresses, devices, orders, cart).
  responses:
    Unauthorized:
      description: Missing, invalid, or expired session/token.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  schemas:
    ErrorResponse:
      type: object
      description: The uniform shape every `apperr.AppError` serializes to.
      properties:
        code: { type: string, example: invalid_phone }
        message: { type: string, example: "укажите номер в формате +996XXXXXXXXX" }
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        name_ru: { type: string }
        name_ky: { type: string }
        description_ru: { type: string, nullable: true }
        description_ky: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        base_price: { type: number, format: double }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Variant:
      type: object
      properties:
        id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        size: { type: string }
        color: { type: string }
        sku: { type: string, nullable: true }
        price_override: { type: number, format: double, nullable: true }
...
    StockPoint:
      type: object
      properties:
        point_id: { type: string, format: uuid }
        quantity: { type: integer }
...
    ProductDetailResponse:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            images:
              type: array
              items: { $ref: "#/components/schemas/ProductPhoto" }
            variants:
              type: array
              items: { $ref: "#/components/schemas/VariantDetail" }
    ProductPhoto:
      type: object
      properties:
        url: { type: string, description: "Full-size (1200×1200) photo." }
        thumb_url: { type: string, description: "400×400 variant, e.g. for a thumbnail strip. Same as url for legacy photos without variants." }
        sort_order: { type: integer }
        color:
          type: string
          nullable: true
          description: "Null for a general product photo. When set, the client should prefer photos whose color matches the shopper's selected variant color, falling back to the null-color photos if that color has none."
    VariantDetail:
      allOf:
        - $ref: "#/components/schemas/Variant"
        - type: object
          properties:
            stock:
              type: array
              items: { $ref: "#/components/schemas/StockPoint" }
...
    CartLine:
      type: object
      description: One line of a customer's server-side cart, enriched with the current variant/product/primary-photo data (GET /api/v1/cart).
      properties:
        variant_id: { type: string, format: uuid }
        quantity: { type: integer }
        product_id: { type: string, format: uuid }
        product_name: { type: string }
        product_name_ky: { type: string }
        size: { type: string }
        color: { type: string }
        price: { type: number, format: double, description: "variant.price_override if set, else the product's base_price." }
        photo_url: { type: string, nullable: true, description: "Primary photo URL, or null if the product has none." }
        thumb_url: { type: string, nullable: true, description: "400×400 variant of the primary photo (same as photo_url for legacy photos), or null if the product has none." }
        is_active: { type: boolean, description: "false once the product was deactivated — the line can't be ordered." }
        in_stock: { type: integer, description: "Most pairs of this variant any one active store holds." }
        available: { type: boolean, description: "is_active and in_stock >= quantity — false means checkout would reject this line." }
...